Created Releases directory, added Upcoming/Past/CurrentComicsList.js & Comic.js. Modified .gitignore file to add new dir to project.
