+++ .gitignore
@@ -20,7 +20,7 @@ mono_crash.*
 [Dd]ebug/
 [Dd]ebugPublic/
 [Rr]elease/
-[Rr]eleases/
+# [Rr]eleases/
 x64/
 x86/
 [Ww][Ii][Nn]32/
