+++ frontend/src/components/Releases/CurrentComicsList.js
@@ -1,18 +1,17 @@
 import React, { useContext, useEffect } from "react";
 // import { useHistory } from "react-router-dom"
-import { ComicContext } from "../providers/ComicProvider";
 import { Section, Container } from "react-bulma-components";
-import { ReleaseDate, WeekStart } from "./UtilityMethods";
+import { ComicContext } from "../../providers/ComicProvider";
+import { ReleaseDate, WeekStart } from "../UtilityMethods";
 import Comic from "./Comic"
-import "../index.css"
 
 
-const UpcomingComicsList = () => {
+const CurrentComicsList = () => {
     // 
     // -- GitHub Issue Ticket # 4 ---
     // [Releases [Ticket #4]](https://github.com/jon-xo/bifrost/issues/5)
     // 
-    //  - newComics and getComics are accessed via ComicContext in the provider
+    //  - currentComics and getCurrentComics are accessed via ComicContext in the provider
     // - Array returned by API is stored in newComics
     // - ReleaseDate & WeekStart are imported from local UtilityMethods,
     //   newComicDay stores a single date as string, if newComicDay
@@ -22,24 +21,25 @@ const UpcomingComicsList = () => {
     //   comic cards.
     //   
     
-    const { newComics, getNewComics } = useContext(ComicContext);
+    const { currentComics, getCurrentComics } = useContext(ComicContext);
 
     useEffect(() => {
-        getNewComics();
-    }, [getNewComics])
+        getCurrentComics();
+    }, [currentComics, getCurrentComics])
 
-    const newComicDay = ReleaseDate(newComics);
+    const newComicDay = ReleaseDate(currentComics);
 
     return (
             <Section>
-                <h2 className="title is-2">Upcoming Comics</h2>
+                <h2 className="title is-2">Current Comics</h2>
             {newComicDay === undefined ? 
                 <h4 className="subtitle is-5">Release Date: ...</h4>
                 :
                 <h4 className="subtitle is-5"><strong>Release Date</strong>: {WeekStart(newComicDay, "mid")}</h4>
             }
                 <Container fluid='true' className='comic-container'>
-                {newComics.comics?.map((comic) => (                
+
+                {currentComics.comics?.map((comic) => (
                     <Comic key={comic.diamond_id} comic={comic} />
                 ))}
                 </Container>
@@ -47,4 +47,4 @@ const UpcomingComicsList = () => {
     ); 
 };
 
-export default UpcomingComicsList;
+export default CurrentComicsList;
